+++ autogpt_platform/backend/pyproject.toml
@@ -46,6 +46,7 @@ youtube-transcript-api = "^0.6.2"
 googlemaps = "^4.10.0"
 replicate = "^0.34.1"
 
+pinecone = "^5.3.1"
 [tool.poetry.group.dev.dependencies]
 poethepoet = "^0.26.1"
 httpx = "^0.27.0"
